requirements.txt
python-telegram-bot>=21.4,<22.0
python-dotenv>=1.0
PyYAML>=6.0
